+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [0.10.0] - 2022-06-14
+# Added
+* Firebase functions
+* Firebase create user function
+* Strapi employee and admin lifecycle webhook
+
 # [0.9.0] - 2022-06-10
 ### Added
 * Implemented all Employee UI & logic: Notifications, User Reservations, Reservation Create.